requirements.txt
requests
pytest-html
webdriver-manager
selenium==4.17.0
pytest==7.4.3
allure-pytest==2.13.2